# gamestore-service/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /tshirts:
    get:
      tags:
        - t-shirt-controller
      operationId: getAllTShirts
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TShirt'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    put:
      tags:
        - t-shirt-controller
      operationId: updateTShirt
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TShirt'
        required: true
      responses:
        '204':
          description: No Content
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    post:
      tags:
        - t-shirt-controller
      operationId: addTShirt
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TShirt'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TShirt'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /tax:
    put:
      tags:
        - tax-controller
      operationId: updateTax
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tax'
        required: true
      responses:
        '204':
          description: No Content
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Tax'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    post:
      tags:
        - tax-controller
      operationId: addTax
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tax'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Tax'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /game:
    get:
      tags:
        - game-controller
      operationId: getAllGames
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    put:
      tags:
        - game-controller
      operationId: updateGame
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Game'
        required: true
      responses:
        '204':
          description: No Content
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Game'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    post:
      tags:
        - game-controller
      operationId: addGame
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Game'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Game'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /fee:
    put:
      tags:
        - fee-controller
      operationId: updateFee
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Fee'
        required: true
      responses:
        '204':
          description: No Content
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Fee'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    post:
      tags:
        - fee-controller
      operationId: addFee
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Fee'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Fee'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /console/{id}:
    get:
      tags:
        - console-controller
      operationId: getConsoleById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Console'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    put:
      tags:
        - console-controller
      operationId: updateConsoleById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Console'
        required: true
      responses:
        '204':
          description: No Content
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    delete:
      tags:
        - console-controller
      operationId: shouldDeleteConsole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /invoice:
    get:
      tags:
        - invoice-controller
      operationId: getAllInvoices
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvoiceViewModel'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    post:
      tags:
        - invoice-controller
      operationId: createInvoice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceViewModel'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InvoiceViewModel'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /console:
    get:
      tags:
        - console-controller
      operationId: getAllConsoles
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Console'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    post:
      tags:
        - console-controller
      operationId: createConsole
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Console'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Console'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /tshirts/{id}:
    get:
      tags:
        - t-shirt-controller
      operationId: getTShirtById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TShirt'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    delete:
      tags:
        - t-shirt-controller
      operationId: deleteTShirtById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /tshirts/size/{size}:
    get:
      tags:
        - t-shirt-controller
      operationId: getTShirtBySize
      parameters:
        - name: size
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TShirt'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /tshirts/color/{color}:
    get:
      tags:
        - t-shirt-controller
      operationId: getTShirtByColor
      parameters:
        - name: color
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TShirt'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /tax/{state}:
    get:
      tags:
        - tax-controller
      operationId: getTaxByState
      parameters:
        - name: state
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Tax'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    delete:
      tags:
        - tax-controller
      operationId: deleteTaxByState
      parameters:
        - name: state
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /invoice/{id}:
    get:
      tags:
        - invoice-controller
      operationId: getInvoiceById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InvoiceViewModel'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /invoice/name/{name}:
    get:
      tags:
        - invoice-controller
      operationId: findInvoicesByCustomerName
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvoiceViewModel'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /game/{id}:
    get:
      tags:
        - game-controller
      operationId: getGameById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Game'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    delete:
      tags:
        - game-controller
      operationId: deleteGameById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /game/title/{title}:
    get:
      tags:
        - game-controller
      operationId: findGamesByTitle
      parameters:
        - name: title
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /game/studio/{studio}:
    get:
      tags:
        - game-controller
      operationId: findGamesByStudio
      parameters:
        - name: studio
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /game/esrbrating/{esrbRating}:
    get:
      tags:
        - game-controller
      operationId: findGamesByEsrbRating
      parameters:
        - name: esrbRating
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /fee/{product_type}:
    get:
      tags:
        - fee-controller
      operationId: getFeeByProductType
      parameters:
        - name: product_type
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Fee'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    delete:
      tags:
        - fee-controller
      operationId: deleteFeeByProductType
      parameters:
        - name: product_type
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /console/manufacturer/{manufacturer}:
    get:
      tags:
        - console-controller
      operationId: shouldGetConsoleByManufacturer
      parameters:
        - name: manufacturer
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Console'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
components:
  schemas:
    CustomErrorResponse:
      type: object
      properties:
        errorMsg:
          type: string
        status:
          type: integer
          format: int32
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time
    TShirt:
      required:
        - color
        - description
        - price
        - quantity
        - size
      type: object
      properties:
        gettShirtId:
          type: integer
          format: int32
        size:
          maxLength: 20
          minLength: 0
          type: string
        color:
          maxLength: 20
          minLength: 0
          type: string
        description:
          maxLength: 255
          minLength: 0
          type: string
        price:
          minimum: 0
          exclusiveMinimum: true
          type: number
        quantity:
          minimum: 1
          type: integer
          format: int32
    Tax:
      required:
        - rate
        - state
      type: object
      properties:
        state:
          maxLength: 2
          minLength: 2
          type: string
        rate:
          minimum: 0
          exclusiveMinimum: true
          type: number
    Game:
      required:
        - description
        - esrbRating
        - price
        - quantity
        - studio
        - title
      type: object
      properties:
        id:
          type: integer
          format: int32
        title:
          maxLength: 50
          minLength: 0
          type: string
        esrbRating:
          maxLength: 50
          minLength: 0
          type: string
        description:
          maxLength: 255
          minLength: 0
          type: string
        price:
          minimum: 0
          exclusiveMinimum: true
          type: number
        studio:
          maxLength: 20
          minLength: 0
          type: string
        quantity:
          minimum: 1
          type: integer
          format: int32
    Fee:
      required:
        - fee
        - productType
      type: object
      properties:
        productType:
          maxLength: 50
          minLength: 0
          type: string
        fee:
          minimum: 0
          exclusiveMinimum: true
          type: number
    Console:
      required:
        - manufacturer
        - memoryAmount
        - model
        - price
        - processor
        - quantity
      type: object
      properties:
        id:
          type: integer
          format: int32
        model:
          maxLength: 50
          minLength: 0
          type: string
        manufacturer:
          maxLength: 50
          minLength: 0
          type: string
        memoryAmount:
          maxLength: 20
          minLength: 0
          type: string
        processor:
          maxLength: 20
          minLength: 0
          type: string
        price:
          minimum: 0
          exclusiveMinimum: true
          type: number
        quantity:
          minimum: 1
          type: integer
          format: int32
    InvoiceViewModel:
      required:
        - city
        - itemId
        - itemType
        - name
        - quantity
        - state
        - street
        - zipcode
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          maxLength: 50
          minLength: 0
          type: string
        street:
          maxLength: 100
          minLength: 0
          type: string
        city:
          maxLength: 50
          minLength: 0
          type: string
        state:
          maxLength: 20
          minLength: 0
          type: string
        zipcode:
          maxLength: 10
          minLength: 0
          type: string
        itemType:
          maxLength: 50
          minLength: 0
          type: string
        itemId:
          type: integer
          format: int32
        unitPrice:
          minimum: 0
          exclusiveMinimum: true
          type: number
        quantity:
          minimum: 1
          type: integer
          format: int32
        subtotal:
          minimum: 0
          exclusiveMinimum: true
          type: number
        tax:
          minimum: 0
          exclusiveMinimum: true
          type: number
        processingFee:
          minimum: 0
          exclusiveMinimum: true
          type: number
        total:
          minimum: 0
          exclusiveMinimum: true
          type: number